move MiLightButtons->MiLightConstants
